+++ src/Tool/Service/ScoreServiceClient.php
@@ -22,25 +22,25 @@ declare(strict_types=1);
 
 namespace OAT\Library\Lti1p3Ags\Tool\Service;
 
+use InvalidArgumentException;
 use OAT\Library\Lti1p3Ags\Model\Score;
 use OAT\Library\Lti1p3Ags\Serializer\Normalizer\Tool\ScorePublishNormalizer;
 use OAT\Library\Lti1p3Core\Message\Claim\AgsClaim;
 use OAT\Library\Lti1p3Core\Registration\RegistrationInterface;
-use OAT\Library\Lti1p3Core\Service\Client\ServiceClient;
+use OAT\Library\Lti1p3Core\Service\Client\ServiceClientInterface;
 use Psr\Http\Message\ResponseInterface;
-use Symfony\Component\Translation\Exception\LogicException;
 
-class ScorePublishService
+class ScoreServiceClient
 {
     public const AUTHORIZATION_SCOPE_SCORE = 'https://purl.imsglobal.org/spec/lti-ags/scope/score';
 
     /** @var ScorePublishNormalizer */
     private $scorePublishNormalizer;
 
-    /** @var ServiceClient */
+    /** @var ServiceClientInterface */
     private $serviceClient;
 
-    public function __construct(ScorePublishNormalizer $scorePublishNormalizer, ServiceClient $serviceClient)
+    public function __construct(ScorePublishNormalizer $scorePublishNormalizer, ServiceClientInterface $serviceClient)
     {
         $this->scorePublishNormalizer = $scorePublishNormalizer;
         $this->serviceClient = $serviceClient;
@@ -49,7 +49,7 @@ class ScorePublishService
     public function publish(RegistrationInterface $registration, AgsClaim $agsClaim, Score $score): ResponseInterface
     {
         if (null === $agsClaim->getLineItemUrl()) {
-            throw new LogicException('The line item url required to send the score is not defined');
+            throw new InvalidArgumentException('The line item url required to send the score is not defined');
         }
 
         return $this->serviceClient->request(